refactor(backend): extract direct prompt tool context
- move provider-aware direct tool binding into direct_prompt_tool_binding.py
- move direct and Code Studio tool-context prompt builders into direct_prompt_tool_context.py
- make skill injection inputs explicit and add focused regression coverage
- update cleanup audit and codebase navigation docs
